use c99 int types instead of custom types + add linker script back

+++ libc/mem.c
@@ -8,7 +8,7 @@
  *
  * Copies n bytes of data from one char to another
  */
-void memory_copy(u8 *source, u8 *dest, int nbytes) {
+void memory_copy(uint8_t *source, uint8_t *dest, int nbytes) {
   int i;
   for (i = 0; i < nbytes; i++) {
     *(dest + i) = *(source + i);
@@ -20,31 +20,30 @@ void memory_copy(u8 *source, u8 *dest, int nbytes) {
  *
  * Set memory address
  */
-void memory_set(u8 *dest, u8 val, u32 len) {
-  u8 *temp = (u8 *)dest;
+void memory_set(uint8_t *dest, uint8_t val, uint32_t len) {
+  uint8_t *temp = (uint8_t *)dest;
   for ( ; len != 0; len--) *temp++ = val;
 }
 
 // @TODO: compute at link time
 // For now we use the hardcoded value we defined in the Makefile
-u32 free_mem_addr = 0x10000;
+uint32_t free_mem_addr = 0x10000;
 
 /**
  * kmalloc
  *
  * Implementation is just a pointer to some free memory which keeps growing
  */
-u32 kmalloc(u32 size, int align, u32 *phys_addr) {
-  // Pages are aligned to 4k, or 0x1000
+uint32_t kmalloc(size_t size, int align, uint32_t *phys_addr) {
+  /* Pages are aligned to 4K, or 0x1000 */
   if (align == 1 && (free_mem_addr & 0xFFFFF000)) {
     free_mem_addr &= 0xFFFFF000;
     free_mem_addr += 0x1000;
   }
-
-  // Save the physical address
+  /* Save also the physical address */
   if (phys_addr) *phys_addr = free_mem_addr;
 
-  u32 ret = free_mem_addr;
+  uint32_t ret = free_mem_addr;
   free_mem_addr += size;
   return ret;
 }
